+++ pixply/lib/Settings/settinggeneral.dart
@@ -13,7 +13,7 @@ import 'package:pixply/Settings/help.dart';
 import 'package:pixply/Settings/youridea.dart';
 import 'package:pixply/explore/explore.dart';
 import 'package:pixply/Likes/like.dart';
-// import 'package:pixply/Settings/playback_state.dart';
+import 'package:pixply/Settings/playback_state.dart';
 // import 'package:pixply/Settings/solid_bmp.dart';
 // import 'package:pixply/UserAccount/user.dart';
 // import 'package:pixply/Settings/termsconditions.dart';
@@ -144,28 +144,28 @@ void toggleSection(String section) {
 
 
 
-  // Future<void> resetToDefaults() async {
-  //   setState(() {
-  //     brightness = 0.5;
-  //     _selectedRotation = ScreenRotation.degree0;
-  //     selectedColor = Colors.white;
-  //     expandedSections.updateAll((key, value) => false);
-  //   });
+  Future<void> resetToDefaults() async {
+    setState(() {
+      brightness = 0.5;
+      _selectedRotation = ScreenRotation.degree0;
+      selectedColor = Colors.white;
+      expandedSections.updateAll((key, value) => false);
+    });
 
-  //   // Update shared config so external settings also reset app-wide defaults
-  //   DisplayManager.recordLastDisplay(
-  //     path: 'assets/logopixply.png',
-  //     type: DisplayType.image,
-  //   );
-  //   ColorConfig.ledMasterBrightness = 0.5;
-  //   ColorConfig.setColor(Colors.white);
-  //   RotationStore.setRotation(ScreenRotation.degree0);
-  //   PlaybackState.notifyReset();
+    // Update shared config so external settings also reset app-wide defaults
+    DisplayManager.recordLastDisplay(
+      path: 'assets/logopixply.png',
+      type: DisplayType.image,
+    );
+    ColorConfig.ledMasterBrightness = 0.5;
+    ColorConfig.setColor(Colors.white);
+    RotationStore.setRotation(ScreenRotation.degree0);
+    PlaybackState.notifyReset();
 
-  //   if (widget.isConnected) {
-  //     await DisplayManager.refreshDisplay(clearBeforeSend: true);
-  //   }
-  // }
+    if (widget.isConnected) {
+      await DisplayManager.refreshDisplay(clearBeforeSend: true);
+    }
+  }
   
 
 
@@ -636,13 +636,13 @@ onChanged: (value) async {
               
             ),
 
-            // _buildTile(
-            //   "Reset to Default",
-            //   resetToDefaults,
-            //   iconPath: 'assets/reset.svg',
+            _buildTile(
+              "Reset to Default",
+              resetToDefaults,
+              iconPath: 'assets/reset.svg',
 
               
-            // ),
+            ),
 _buildTile(
   "Help Center",
   () {